requirements.txt
Django==6.0.1
djangorestframework==3.16.1
drf-spectacular==0.29.0
django-filter==25.2
black==25.12.0